Configuring pass as LLVM project.
You should now run ./configure before using this project and set the
LLVM_CONFIG environmental variable to the path of the llvm-config
executable of the installation you want to use. Running "make" will
compile the project and "make install" will install the generated shared
libraries in the <build-type>/lib directory of your LLVM build.
This should work for LLVM versions 3.4 to 3.7, possibly others.